implement clean architecture
