+++ lib/features/home/presentation/pages/home_page.dart
@@ -1,0 +1,54 @@
+import 'package:demo/shared/assets/app_images.dart';
+import 'package:demo/features/home/presentation/bloc/counter_cubit.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+
+class HomePage extends StatefulWidget {
+  const HomePage({super.key});
+
+  @override
+  State<HomePage> createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  @override
+  Widget build(BuildContext context) {
+    final counterBlocprovider = BlocProvider.of<CounterCubit>(context);
+    return Scaffold(
+      appBar: AppBar(
+        title: Image.asset(AppImages.logo, height: 80, fit: BoxFit.cover),
+        centerTitle: true,
+        toolbarHeight: 100,
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text('You have pushed the button this many times:'),
+            BlocBuilder<CounterCubit, int>(
+              bloc: counterBlocprovider,
+              builder: (context, counter) {
+                return Text(
+                  '$counter',
+                  style: Theme.of(context).textTheme.headlineMedium,
+                );
+              },
+            ),
+            ElevatedButton(
+              onPressed: () => context.push("/nextPage"),
+              child: Text('Next Page'),
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {
+          counterBlocprovider.increment();
+        },
+        tooltip: 'Increment',
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
