+++ lib/shared/assets/app_images.dart
@@ -1,0 +1,9 @@
+mixin AppImages {
+  static const String logo = 'assets/images/logo.png';
+  static const String splash = 'assets/images/splash_screen.png';
+  // static const String noData = 'assets/images/no_data.png';
+  // static const String noInternet = 'assets/images/no_internet.png';
+  // static const String error = 'assets/images/error.png';
+  // static const String empty = 'assets/images/empty.png';
+  // static const String emptyCart = 'assets/images/empty_cart.png';
+}
